requirements.txt
aiogram==3.21.0
openai==1.99.9

fastapi==0.115.4
uvicorn[standard]==0.30.3
httpx==0.28.1
anyio==4.10.0
h11==0.16.0
sniffio==1.3.1

aiohttp==3.12.15
magic-filter==1.0.12
yarl==1.20.1
attrs==25.3.0
frozenlist==1.7.0
aiosignal==1.4.0

motor==3.6.0
pymongo==4.9.1

python-dotenv==1.0.1
tqdm==4.67.1
distro==1.9.0
typing_extensions==4.14.1
pydantic==2.11.7
pydantic_core==2.33.2
jiter==0.10.0
propcache==0.3.2
typing-inspection==0.4.1

pydub>=0.25.1
reportlab>=4.2.0
